Move `SpeculativeConfig` from `config/__init__.py` to `config/speculative.py`
