+++ tests/v1/spec_decode/test_ngram.py
@@ -2,7 +2,8 @@
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 import numpy as np
 
-from vllm.config import ModelConfig, SpeculativeConfig, VllmConfig
+from vllm.config import ModelConfig, VllmConfig
+from vllm.config.speculative import SpeculativeConfig
 from vllm.v1.spec_decode.ngram_proposer import (
     NgramProposer, _find_longest_matched_ngram_and_propose_tokens)
